extract common helper methods from metric capture
